Add user auth, JWT security, Flyway migrations, and full DB persistence
- User registration and login with BCrypt password hashing and JWT tokens
- Spring Security with stateless JWT filter protecting cart and order endpoints
- Converted Product and Order from in-memory to JPA entities persisted in H2
- Order now has one-to-many relationship with OrderItem (proper cart checkout)
- Orders linked to the authenticated user who placed them
- Flyway migrations (V1-V4) replace ddl-auto — schema changes now versioned
- H2 switched from in-memory to file-based so data survives restarts
- Frontend: Register/Login pages, AuthContext, protected Store route, logout
- Frontend: authFetch utility adds JWT Authorization header automatically
- Removed Student entity, Explorer page — focused on Quickmart domain

+++ src/main/java/org/example/controller/ProductController.java
@@ -28,7 +28,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable int id) {
+    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
         log.info("→ [ProductController] GET /api/products/{}", id);
         ResponseEntity<Product> response = productService.getProductById(id)
                 .map(p -> {
@@ -51,7 +51,7 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable int id, @RequestBody Product product) {
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product product) {
         log.info("→ [ProductController] PUT /api/products/{}, new name='{}', price={}", id, product.getName(), product.getPrice());
         ResponseEntity<Product> response = productService.updateProduct(id, product)
                 .map(p -> {
@@ -66,7 +66,7 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable int id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
         log.info("→ [ProductController] DELETE /api/products/{}", id);
         boolean deleted = productService.deleteProduct(id);
         if (deleted) {
